OVC5 parallel ethernet interfaces (#82)
* Skeleton of imager assignment to multiple sockets


* Skeleton host driver for multiple ethernets


* Add second client to device


* Add second subscriber to server


* WIP multiple ethernet interfaces streaming


* WIP remove spinlock


* Roll back to waiting whole frame before send

* Add byteswap function for RAW16 images


* Add greyscale image


* Add MLX module


* Initial support for ToF frames


* Increase clock speed / redistribute DDR interfaces


* Add deep copy to returned frames


* WIP temporarily drop framerate and bypass interrupts for polling


* Back to main camera setup, increase clock speed


* First full integration demo with AR0234 hack


* Refactor to parallel interrupt polling / socket send


* Fully multithreaded interrupt polling


* Multithreaded server and save float images


* Streamline clocks, enforce dma coherency


* Add USB hub and IRQ affinity


* Add frame id to publisher


* Add interrupt controller, free up GIC interrupts


* Minor cleanup

* Upgrade to Debian Bullseye


* Make script executable


* Fix MTU, update rc.local


* Inline rclocal writing

+++ ovc5/software/camera_device_driver/include/ovc5_driver/camera.hpp
@@ -14,10 +14,10 @@ typedef struct camera_dynamic_configs_t
 
 typedef struct camera_params_t
 {
-  int res_x = -1;
-  int res_y = -1;
-  int fps = -1;
-  int bit_depth = -1;
+  long res_x = -1;
+  long res_y = -1;
+  long fps = -1;
+  long bit_depth = -1;
   char data_type[8];
   camera_dynamic_configs_t dynamic_configs = {0};
 } camera_params_t;
